[STYLE] 커뮤니티 컴포넌트 분리, 로딩 추가
