+++ src/pages/community/Community.styles.ts
@@ -5,52 +5,7 @@ export const CommunityContainer = styled.div`
     align-items: center;
     flex-direction: column;
 `
-// 프로필
-export const PostProFileBox = styled.div`
-    width: 330px;
-    height: 60px;
-    background-color: #F5FAF1;
-    display: flex;
-    padding: 5px;
-    border-radius: 10px;
-    
-`
 
-export const PostProFileImgBox = styled.div`
-    width: 48px;
-    height: 48px;
-    border-radius: 100px;
-    background-color: #fff;
-    display: flex;
-    justify-content: center; 
-    align-items: center;
-    margin-right: 10px;
-`
-export const PostProFileTitle = styled.p`
-    font-weight: bold;
-    font-size: 10px;
-`
-
-
-// 글쓰기
-export const PostButtonBox = styled.div`
-    width: 321px;
-    height: 20px;
-    margin: 15px;
-    display: flex;
-`
-
-export const CompositionButton =styled.button`
-    width: 72px;
-    height: 20px;
-    background-color: #72A474;
-    color: #fff;
-    font-size: 12px;
-    display: flex;
-    justify-content: center; 
-    align-items: center;
-    
-`
 
 // 카드
 export const PostContainerScroll = styled.div`
