+++ src/pages/mypage/MyComponent.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { SharedInput } from '../Shared.styles';
 import { MyComponentButton, MyComponentContainer } from './MyComponent.styles';
-import { diariesApi } from '../../api/http';
+import { myplantApi } from '../../api/http';
 
 
 const MyComponent: React.FC = () => {
@@ -21,7 +21,7 @@ const MyComponent: React.FC = () => {
     // 수정을 저장할 때 PATCH 요청을 보냅니다.
     const saveChanges = async () => {
         try {
-            const response = await diariesApi.patch(`/api/random/templates/${templateId}`, {
+            const response = await myplantApi.patch(`/api/random/templates/${templateId}`, {
                 answer: editedEntry, 
             });
             navigate('/myplant');
