+++ src/route/Routes.tsx
@@ -8,6 +8,7 @@ import {
 import Cookies from 'js-cookie';
 import Layout from '../components/atoms/Layout';
 import { GlobalStyles } from '../components/atoms/GlobalStyled';
+import Spinner from '/Spin.gif'
 
 // Lazy-loaded components
 const SignIn = lazy(() => import('../pages/login/signin/SignIn'));
@@ -18,7 +19,7 @@ const MyPlantPage = lazy(() => import('../pages/mypage/MyPlantPage'));
 const MyComponent = lazy(() => import('../pages/mypage/MyComponent'));
 const PlantRegistrationPage = lazy(() => import('../pages/mypage/registration/PlantRegistrationPage'));
 const CommunityPage = lazy(() => import('../pages/community/CommunityPage'));
-const CommunityWritePage = lazy(() => import('../pages/community/CommunityWritePage'));
+const CommunityWritePage = lazy(() => import('../pages/community/write/CommunityWritePage'));
 const MyProfile = lazy(() => import('../pages/myprofile/MyProfilepage'));
 const MainPage = lazy(() => import('../pages/mainpage/MainPage'));
 const BoardTest = lazy(() => import('../pages/boardtest/boardtestpage/BoardTest'));
@@ -53,7 +54,7 @@ export const Routes = () => {
     return (
         <Router>
             <GlobalStyles />
-            <Suspense fallback={<div>loading...</div>}>
+            <Suspense fallback={<div  style={{ position: 'fixed', top: '50%', left: '50%', transform: 'translate(-50%, -50%)' }}> <img src={Spinner} alt="Loding" /></div>}>
                 <ReactRouterRoutes>
                     <Route element={<Layout />}>
                         {RouterInfo.map(({ path, element, withAuthorization }) => (
